--- .changeset/crypto-vrf-sign.md
@@ -1,5 +1,0 @@
----
-"@talismn/crypto": minor
----
-
-Add sr25519SignVrf / sr25519VerifyVrf for sr25519 VRF signatures in the origin-bound substrate-vrf context namespace
