+++ .changeset/substrate-vrf-initial.md
@@ -1,0 +1,5 @@
+---
+"@talismn/substrate-vrf": major
+---
+
+New MIT package: sr25519 VRF sign/verify in the origin-bound substrate-vrf context namespace
