+++ apps/extension/src/core/handlers/Extension.spec.ts
@@ -3,9 +3,9 @@ import { getSs58AddressInfo } from "@polkadot-api/substrate-bindings"
 import { mergeUint8 } from "@polkadot-api/utils"
 import { base64 } from "@scure/base"
 import { verify as sr25519Verify } from "@scure/sr25519"
-import { sr25519VerifyVrf } from "@talismn/crypto"
 import type { Account } from "@talismn/keyring"
 import { CUSTOM_SIGNED_EXTENSIONS, getPjsTxHelper } from "@talismn/sapi"
+import { sr25519VerifyVrf } from "@talismn/substrate-vrf"
 import { hexToU8a, u8aToHex } from "@talismn/util"
 import { waitFor } from "@testing-library/dom"
 import { v4 } from "uuid"
