+++ packages/crypto/src/index.ts
@@ -5,14 +5,6 @@ export * from "./hashing"
 export * from "./keystore"
 export * from "./mnemonic"
 export * from "./platform"
-// signing exposes only the wallet/dapp surface: raw schnorrkel VRF primitives and the
-// namespace builder stay internal (see their docstrings)
-export {
-  SIGNATURE_TYPE_PREFIX,
-  type SubstrateVrfNamespace,
-  signSubstrate,
-  sr25519SignVrf,
-  sr25519VerifyVrf,
-} from "./signing"
+export * from "./signing"
 export * from "./types"
 export * from "./utils"
