+++ packages/substrate-vrf/src/index.ts
@@ -1,0 +1,6 @@
+export {
+  type SubstrateVrfNamespace,
+  sr25519SignVrf,
+  sr25519VerifyVrf,
+  substrateVrfContext,
+} from "./vrf"
