+++ packages/substrate-vrf/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@talismn/tsconfig/es6-node.json",
+  "include": ["src"],
+  "exclude": ["**/node_modules", "**/.*/"]
+}
