+++ packages/substrate-vrf/tsdown.config.mts
@@ -1,0 +1,11 @@
+import { defineConfig } from "tsdown"
+
+export default defineConfig({
+  entry: ["src/index.ts"],
+  format: ["cjs", "esm"],
+  dts: true,
+  sourcemap: true,
+  target: "es2022",
+  outExtensions: ({ format }) =>
+    format === "cjs" ? { js: ".js", dts: ".d.ts" } : { js: ".mjs", dts: ".d.mts" },
+})
